requirements.txt
flask==3.0.0
flask_sqlalchemy==3.1.1
flask_login==0.6.3
flask_wtf==1.2.1
email_validator==2.2.0
python-dotenv==1.0.1
